# solution/docs/swagger.yaml
components:
  schemas:
...
    dto.DailyStats:
      properties:
        clicks_count:
          type: integer
        conversion:
          type: number
        date:
          type: integer
        impressions_count:
          type: integer
        spent_clicks:
          type: number
        spent_impressions:
          type: number
        spent_total:
          type: number
      type: object
...
    dto.Stats:
      properties:
        clicks_count:
          type: integer
        conversion:
          type: number
        impressions_count:
          type: integer
        spent_clicks:
          type: number
        spent_impressions:
          type: number
        spent_total:
          type: number
      type: object
    dto.TimeSetBody:
      properties:
        current_date:
          type: integer
      required:
      - current_date
      type: object
...
    models.MLScore:
      properties:
        advertiser_id:
          type: string
        client_id:
          description: составной первичный ключ
          type: string
        score:
          type: integer
      required:
      - advertiser_id
      - client_id
      - score
      type: object
...
info:
  description: PROOOOOOOOOD
  title: Prod-2025 AD API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /ml-scores:
    post:
      description: Set the ML score for a specific entities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.MLScore'
        description: ML Score
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Set ML score
      tags:
      - system
...
  /stats/advertisers/{id}/campaigns:
    get:
      description: Get statistics for all campaigns of a specific advertiser by advertiser
        ID
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Stats'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get advertiser campaign statistics
      tags:
      - stats
  /stats/advertisers/{id}/campaigns/daily:
    get:
      description: Get daily statistics for all campaigns of a specific advertiser
        by advertiser ID
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.DailyStats'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get advertiser campaign daily statistics
      tags:
      - stats
  /stats/campaigns/{id}:
    get:
      description: Get statistics for a specific campaign by campaign ID
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Stats'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get campaign statistics
      tags:
      - stats
  /stats/campaigns/{id}/daily:
    get:
      description: Get daily statistics for a specific campaign by campaign ID
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.DailyStats'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get campaign daily statistics
      tags:
      - stats
  /time/advance:
    post:
      description: Advance the current date in the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.TimeSetBody'
        description: Time Set Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.TimeSetBody'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Advance time
      tags:
      - system
servers:
- url: http://localhost:8080/